Update to v099r15

+++ nSide/processor/r65816/memory.hpp
@@ -20,7 +20,7 @@ alwaysinline auto readLong(uint addr) -> uint8 {
 }
 
 alwaysinline auto readDB(uint addr) -> uint8 {
-  return read(r.db << 16 | uint16(addr));
+  return read(uint24(r.db << 16) | addr);
 }
 
 alwaysinline auto readPB(uint addr) -> uint8 {
@@ -61,7 +61,7 @@ alwaysinline auto writeLong(uint addr, uint8 data) -> void {
 }
 
 alwaysinline auto writeDB(uint addr, uint8 data) -> void {
-  write(r.db << 16 | uint16(addr), data);
+  write(uint24(r.db << 16) | addr, data);
 }
 
 alwaysinline auto writePB(uint addr, uint8 data) -> void {
